Отрефакторил бизнес логику, навёл порядок, первый этап

+++ data/src/main/java/com/learning/lessons/data/api/question_group/ApiAnswer.kt
@@ -2,7 +2,7 @@ package com.learning.lessons.data.api.question_group
 
 data class ApiAnswer(
     val active: Boolean = false,
-    val correct: Long = 0,
+    val correct: Int = 0,
     val optional_image_url: String = "",
     val text: String = ""
 )
